% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{format_pair_resp_to_df}
\alias{format_pair_resp_to_df}
\title{Format response from retrieval of paired end sequences}
\usage{
format_pair_resp_to_df(resps)
}
\arguments{
\item{resps}{A list of responses from an IRIDA API call}
}
\description{
Format response from retrieval of paired end sequences
}
