% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{format_sequence_resp_to_df}
\alias{format_sequence_resp_to_df}
\title{Format output from retrieval of sequences}
\usage{
format_sequence_resp_to_df(resps)
}
\arguments{
\item{resps}{A list of responses from an IRIDA API call}
}
\description{
Format output from retrieval of sequences
}
