% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_all_sequence_info}
\alias{get_all_sequence_info}
\title{Get sequences associated with a list of IRIDA samples}
\usage{
get_all_sequence_info(samples, n_con = 10)
}
\arguments{
\item{samples}{A list of IRIDA sample IDs}

\item{n_con}{How many connections to send to IRIDA in parralel?}
}
\value{
dataframe of sequences associated with each IRIDA sample
}
\description{
Use this function to get a table of sequences associated with the
supplied IRIDA sample IDs. This function does two API calls per
sample, the first to get ALL sequences, and the second to get more
information on paired sequences. The returned dataframe contains
one record per unique sequence.
}
