% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_forward_and_reverse_files_as_df}
\alias{get_forward_and_reverse_files_as_df}
\title{Retrieve the forward and reverse sequencing file information}
\usage{
get_forward_and_reverse_files_as_df(x)
}
\description{
Retrieve the forward and reverse sequencing file information
}
