% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects.R
\name{get_projects}
\alias{get_projects}
\title{Get all the project information that the user has access to.}
\usage{
get_projects()
}
\description{
Get all the project information that the user has access to.
}
