% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_sample_id_from_api_call}
\alias{get_sample_id_from_api_call}
\title{Get sample ID from the URL of an API call}
\usage{
get_sample_id_from_api_call(url)
}
\description{
Get sample ID from the URL of an API call
}
