% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_sequences}
\alias{get_sequences}
\title{Get sequences associated with a list of IRIDA samples}
\usage{
get_sequences(
  samples,
  n_con = 10,
  type = c("all", "pairs"),
  throttle_capacity = 100,
  fill_time = 10
)
}
\arguments{
\item{samples}{A list of IRIDA sample IDs}

\item{n_con}{Number of connections to send to IRIDA at once.}

\item{type}{all: Retreive all sequence information, pairs: retrieve only paired-read sequences}

\item{throttle_capacity}{From \link[httr2:req_throttle]{httr2::req_throttle}: Set this to limit the number of requests per set fill_time}

\item{fill_time}{From \link[httr2:req_throttle]{httr2::req_throttle}: how much time to fill capacity}
}
\value{
dataframe of sequences associated with each IRIDA sample
}
\description{
Use this function to get a table of sequences associated with the
supplied IRIDA sample IDs. This function does two API calls per
sample, the first to get ALL sequences, and the second to get more
information on paired sequences. The returned dataframe contains
one record per unique sequence.
}
