% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{irida_timecode_to_datetime}
\alias{irida_timecode_to_datetime}
\title{Convert IRIDA timestamps into a datetime}
\usage{
irida_timecode_to_datetime(x)
}
\description{
Convert IRIDA timestamps into a datetime
}
