% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{pair_resps_to_df}
\alias{pair_resps_to_df}
\title{Convert a list of IRIDA API responses for sequence pairs into a dataframe}
\usage{
pair_resps_to_df(resps)
}
\arguments{
\item{resps}{description}
}
\description{
Convert a list of IRIDA API responses for sequence pairs into a dataframe
}
