% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{project_samples}
\alias{project_samples}
\title{Get sample information from all samples in an irida project}
\usage{
project_samples(project_id)
}
\arguments{
\item{project_id}{The IRIDA project ID}
}
\value{
A dataframe with all the samples in a project, along with some
        of the metadata fields I expect to always be filled.
}
\description{
Get sample information from all samples in an irida project
}
