% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{req_irida_sequences}
\alias{req_irida_sequences}
\title{Retrieve sequences related to an IRIDA sample ID}
\usage{
req_irida_sequences(sample_id, type = c("all", "pairs"))
}
\arguments{
\item{sample_id}{IRIDA sample ID}

\item{type}{If all, get all sequences, if pairs, get only paired end reads}
}
\description{
Retrieve sequences related to an IRIDA sample ID
}
