% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{resp_irida_to_dataframe}
\alias{resp_irida_to_dataframe}
\title{Attempts to format IRIDA api responses to a single dataframe}
\usage{
resp_irida_to_dataframe(resps)
}
\arguments{
\item{resps}{A list of responses, ideally named}
}
\description{
Attempts to format IRIDA api responses to a single dataframe
}
